Fix: using inheritance in Driver for getters and setters
